backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# AI/ML Dependencies
tensorflow==2.15.0
tensorflow-hub==0.14.0
nltk==3.8.1
openai==1.3.7
numpy==1.24.3
nlpaug==1.1.11

# Utilities
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
mlflow==2.8.1
pandas==2.0.3

# Development Tools
pytest==7.4.2
black==23.9.1
flake8==6.1.0